refactor: add useUserContactDetails, DynamicHeight on Profile
